jextract thin tool, maximum api - initial refactoring
