+++ src/jdk.incubator.jextract/share/classes/jdk/incubator/jextract/StructBuilder.java
@@ -22,18 +22,16 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package jdk.incubator.jextract.tool;
+package jdk.incubator.jextract;
 
 import jdk.incubator.foreign.MemoryAddress;
 import jdk.incubator.foreign.MemoryLayout;
 
-import java.lang.constant.DirectMethodHandleDesc;
-
 /**
  * This class generates static utilities class for C structs, unions.
  */
-class StructBuilder extends JavaSourceBuilder {
-    StructBuilder(String className, String pkgName, ConstantHelper constantHelper) {
+public class StructBuilder extends JavaSourceBuilder {
+    public StructBuilder(String className, String pkgName, ConstantHelper constantHelper) {
         super(className, pkgName, constantHelper);
     }
