+++ src/jdk.incubator.jextract/share/classes/jdk/incubator/jextract/Utils.java
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package jdk.incubator.jextract.tool;
+package jdk.incubator.jextract;
 
 import javax.lang.model.SourceVersion;
 import javax.tools.JavaFileObject;
